aplicacao/requirements.txt
oracledb==1.3.2
python-dotenv==1.0.0
questionary==1.10.0